split functionality into multiple methods
