+++ .gitignore
@@ -1,1 +1,2 @@
 .venv/
+__pycache__/
